Examples: use '%zu' in the format string for size_t

+++ src/grib_context.c
@@ -84,7 +84,7 @@ static void* default_long_lasting_malloc(const grib_context* c, size_t size)
     void* ret;
     ret = malloc(size);
     if (!ret) {
-        grib_context_log(c, GRIB_LOG_FATAL, "default_long_lasting_malloc: error allocating %lu bytes", (unsigned long)size);
+        grib_context_log(c, GRIB_LOG_FATAL, "default_long_lasting_malloc: error allocating %zu bytes", size);
         Assert(0);
     }
     return ret;
@@ -100,7 +100,7 @@ static void* default_buffer_malloc(const grib_context* c, size_t size)
     void* ret;
     ret = malloc(size);
     if (!ret) {
-        grib_context_log(c, GRIB_LOG_FATAL, "default_buffer_malloc: error allocating %lu bytes", (unsigned long)size);
+        grib_context_log(c, GRIB_LOG_FATAL, "default_buffer_malloc: error allocating %zu bytes", size);
         Assert(0);
     }
     return ret;
@@ -111,7 +111,7 @@ static void* default_buffer_realloc(const grib_context* c, void* p, size_t size)
     void* ret;
     ret = realloc(p, size);
     if (!ret) {
-        grib_context_log(c, GRIB_LOG_FATAL, "default_buffer_realloc: error allocating %lu bytes", (unsigned long)size);
+        grib_context_log(c, GRIB_LOG_FATAL, "default_buffer_realloc: error allocating %zu bytes", size);
         Assert(0);
     }
     return ret;
@@ -127,7 +127,7 @@ static void* default_malloc(const grib_context* c, size_t size)
     void* ret;
     ret = malloc(size);
     if (!ret) {
-        grib_context_log(c, GRIB_LOG_FATAL, "default_malloc: error allocating %lu bytes", (unsigned long)size);
+        grib_context_log(c, GRIB_LOG_FATAL, "default_malloc: error allocating %zu bytes", size);
         Assert(0);
     }
     return ret;
@@ -138,7 +138,7 @@ static void* default_realloc(const grib_context* c, void* p, size_t size)
     void* ret;
     ret = realloc(p, size);
     if (!ret) {
-        grib_context_log(c, GRIB_LOG_FATAL, "default_realloc: error allocating %lu bytes", (unsigned long)size);
+        grib_context_log(c, GRIB_LOG_FATAL, "default_realloc: error allocating %zu bytes", size);
         Assert(0);
     }
     return ret;
@@ -905,7 +905,7 @@ void* grib_context_malloc_persistent(const grib_context* c, size_t size)
     void* p = c->alloc_persistent_mem(c, size);
     if (!p) {
         grib_context_log(c, GRIB_LOG_FATAL,
-                         "grib_context_malloc_persistent: error allocating %lu bytes", (unsigned long)size);
+                         "grib_context_malloc_persistent: error allocating %zu bytes", size);
         Assert(0);
     }
     return p;
@@ -937,7 +937,7 @@ void* grib_context_malloc(const grib_context* c, size_t size)
     else
         p = c->alloc_mem(c, size);
     if (!p) {
-        grib_context_log(c, GRIB_LOG_FATAL, "grib_context_malloc: error allocating %lu bytes", (unsigned long)size);
+        grib_context_log(c, GRIB_LOG_FATAL, "grib_context_malloc: error allocating %zu bytes", size);
         Assert(0);
     }
     return p;
@@ -950,7 +950,7 @@ void* grib_context_realloc(const grib_context* c, void* p, size_t size)
         c = grib_context_get_default();
     q = c->realloc_mem(c, p, size);
     if (!q) {
-        grib_context_log(c, GRIB_LOG_FATAL, "grib_context_realloc: error allocating %lu bytes", (unsigned long)size);
+        grib_context_log(c, GRIB_LOG_FATAL, "grib_context_realloc: error allocating %zu bytes", size);
         return NULL;
     }
     return q;
@@ -985,7 +985,7 @@ void* grib_context_buffer_malloc(const grib_context* c, size_t size)
     else
         p = c->alloc_buffer_mem(c, size);
     if (!p) {
-        grib_context_log(c, GRIB_LOG_FATAL, "grib_context_buffer_malloc: error allocating %lu bytes", (unsigned long)size);
+        grib_context_log(c, GRIB_LOG_FATAL, "grib_context_buffer_malloc: error allocating %zu bytes", size);
         return NULL;
     }
     return p;
@@ -1003,7 +1003,7 @@ void* grib_context_buffer_realloc(const grib_context* c, void* p, size_t size)
 {
     void* q = c->realloc_buffer_mem(c, p, size);
     if (!q) {
-        grib_context_log(c, GRIB_LOG_FATAL, "grib_context_buffer_realloc: error allocating %lu bytes", (unsigned long)size);
+        grib_context_log(c, GRIB_LOG_FATAL, "grib_context_buffer_realloc: error allocating %zu bytes", size);
         return NULL;
     }
     return q;
